Use active_admin_authorization as Ransack auth_object option (#8143)
send auth_object option to ransack

fixes #8108

The issue suggest to use `active_admin_authorization` as auth_object. It
make sense to me.

I'm not implementing the idea of a default implementation of ransackable
attributes. That doesn't seems easy to me. I would appreciate a PR or a
battle tested example for that.