Use eslint.config.js to replace deprecated config
ESLint has had several config approaches but the new one that is feature complete is eslint.config.js and will be the default in v9. The others are deprecated.

Several rules I checked from our original config were already deprecated. Several are just code formatting related but ESLint has deprecated a large list of those already for v8. We'll revisit code formatting with the suggestion they give which is to use prettier.