Remove redundant gherkin-lint ignore file
The ignore file entry was only needed for CI since it used to run in a larger workflow that also handled Ruby based lint but that's no more so this isn't needed. We run all linters individually. This gives us better control over when that workflow should or should not run.