Use min-w-max for dropdown menu list
Still some problems with the width defaulting to the size of the button when considering desktop and mobile viewports. This is the best we can do for now. One consideration is we can apply a custom class for batch actions and (table) actions so we can set a default width. That will render a bit nicer and more inline with how frameworks style dropdowns.