Extract flash_messages to partial with error type
This updates the icons used as well. The notice type now uses a checkmark, the alert type uses an info icon and the error type uses the exclamation mark icon.

Added a flash_error CSS class for error type since used by Devise and a few places in our code base too.

This simplifies the layout spec as it was failing intermittently with the flash_messages partial extraction. Since we only have 2 unit tests for a simple method. We don't need to create a whole admin app. The lang attributes test is safe to remove as we have an integration test (cucumber) to ensure it works as expected.

Updates the page base test to add and remove the necessary constant so we clean up properly.