Prep Tailwind CSS migration part 7 (#8155)
* Update user menu as drop down

Adds new ids for main and user menus.

* Include plugin.js in published npm package

Tested that this file would not be included by default by running `npm publish --dry-run` which lists the files in the tarball contents. After updating the files array and rerunning the command, the file is now included.

* Add default CSS class only if present

This updates our Component base class to only apply a default CSS class name if present.

This also resolves the issue around the switch_index_view feature test so we can re-enable that.

* Update page title bar styles

* Only render all action items if present

If this div element is there an empty, it will affect the flex layout since we use gap property to control spacing at the parent level.

* Use min-w-max for dropdown menu list

Still some problems with the width defaulting to the size of the button when considering desktop and mobile viewports. This is the best we can do for now. One consideration is we can apply a custom class for batch actions and (table) actions so we can set a default width. That will render a bit nicer and more inline with how frameworks style dropdowns.

* Update comments styles

The comments component is now just a plain div component and no longer relies on panel. The styles are further cleaned up.

* Add client side validation for comments form

By default Formtastic disables form client side validation with novalidate="novalidate" boolean attribute, regardless if we set input html with required boolean.

* Extract flash_messages to partial with error type

This updates the icons used as well. The notice type now uses a checkmark, the alert type uses an info icon and the error type uses the exclamation mark icon.

Added a flash_error CSS class for error type since used by Devise and a few places in our code base too.

This simplifies the layout spec as it was failing intermittently with the flash_messages partial extraction. Since we only have 2 unit tests for a simple method. We don't need to create a whole admin app. The lang attributes test is safe to remove as we have an integration test (cucumber) to ensure it works as expected.

Updates the page base test to add and remove the necessary constant so we clean up properly.

* Fix batch actions JS tr click if no checkbox

* Update attributes_table style with bottom spacing

Also adds overflow hidden to for the rounded corners. We'll have to readjust.

* Bump flowbite to 2.2.0 with modal fix

* Require Arbre v2 with empty data attributes

This now requires Arbre v2 as the minimum since a key change was a bad commit around handling empty attributes. We can now use an empty attributes safely as it will be output to HTML as expected. This act as boolean/presence attributes.