Update Dashboard template
The sample content is now inline (no blank slate) with Tailwind classes and Arbre usage to act as a simple/starter demo of what ActiveAdmin can do.

No need to localize the page text since this is always meant to be deleted and customized. The important parts are already localized.