Extract sidebar to partials
This allows users to customize the sidebar further if needed since these are simple components.

The `attributes_table` method has been removed since it was a bad change. Sidebars can be used outside the show page context so by using `resource` internally that can lead to an unexpected error. This cleans up the API and matches what we have in the docs since we consistently show `attributes_table_for object, ...` usage.

This also removes the unused sidebar methods from resource controller. This same module is duplicated and available as view helpers which are public so this is unnecessary. It has no references, nor any documentation. This adds test coverage for the sidebar helper which was missing.