Partial revert of #5163 for main menu extraction
I do think the changes in #5163 are good but ideally the sorting should be done at the data layer, not in the DOM (Arbre). It has too though since the label, url, etc. have to rendered within the view context so its convenient to do there. I did this partial revert, because it would be easiest to use this within an extracted partial that the app host can customize to meet their needs. It is expected that a context has to be passed in for label, etc. since we have to render these blocks so no real way around that. The menu and menu item files haven't changed before #5163 for many years so I think the API is stable and solid for now.