Update devise login with full width button
The original selector using `:has` wouldn't work in current version of Firefox. This was just a clever way to apply the styles in a specific context but was only for Devise forms which don't matter since we expect users to customize them. We can apply a simple default in the template to grow the flex container when its the only action.