Move link style to page content container
Now that we've extracted the site header, page header and site footer, there is no need for a default link style since it's expected that this will be customized by users. We can limit the link style to just the page content container as a default there where it makes most sense. With how its defined, the links should be able to be overridden with Tailwind utilities now.