Remove actions dropdown from index-as-table
This has scrolling issues within a responsive table for mobile support. We'll maintain the actions as-is which is a better default for the admin. Users can already specify custom columns so they can add a dropdown or another custom component if they want without depending on ActiveAdmin for providing it.