Prep Tailwind CSS migration part 8 (#8170)
* Update scopes and index-list styles

* Fix dropdown placement

* Remove unnecessary base font-size/line-height

No need for these on form controls since it already inherits base which is what we want. This way we can control font size from container.

* Remove sidebar attibutes_table styles

We don't use these. Simple enough for users to provide themselves. We'll have a new attributes_list component that uses dl/dt/dd elements which allow for easier styling in general and for smaller viewports.

* Update devise login with full width button

The original selector using `:has` wouldn't work in current version of Firefox. This was just a clever way to apply the styles in a specific context but was only for Devise forms which don't matter since we expect users to customize them. We can apply a simple default in the template to grow the flex container when its the only action.

* Move link style to page content container

Now that we've extracted the site header, page header and site footer, there is no need for a default link style since it's expected that this will be customized by users. We can limit the link style to just the page content container as a default there where it makes most sense. With how its defined, the links should be able to be overridden with Tailwind utilities now.

* Remove actions dropdown from index-as-table

This has scrolling issues within a responsive table for mobile support. We'll maintain the actions as-is which is a better default for the admin. Users can already specify custom columns so they can add a dropdown or another custom component if they want without depending on ActiveAdmin for providing it.

* Update active filters layout and styles

* Complete filter form styles

* Rename to data-table-resource-actions class

* Consolidate panel styles

* Update status tag classes

* Update paginated collection classes

* Remove redundant outer div from index table

* Clean up index-as-table styles

* Update dropdown menu classes

* Rename to batch-actions-dropdown class name

* Set dropdown options locally