Remove unused view factory class/config
This predates Arbre builder method which is how we create components. This may have been used to override views entirely but with partials we provide a simpler, more Rails like way to customize the admin while minimizing the library footprint and maintainance.

Note that we use the method_or_proc_helper throughout but mysteriously this is defined in view helpers but at the root level module. We already require and include the module elsewhere, e.g. menu item and scopes, so requiring it in filters/forms isn't abnormal. With the view factory removal, the test suite starting failing here but now all green.