Extract layout and pages to partials (#8172)
* Add layout helper from base page w/titles

Should be able to handle the site, page and combined titles for both layout files.

Consolidate title helper into layout helper.

* Extract layout file

Known issue with show action when declaring a member_action to render a page, no page title is rendered.

We won't include the body tag data attributes since we don't utilize that and since there's no need now as users can customize.

* Add page title and presenter methods

* Extract page template

* Extract show template

* Extract new/edit form templates

* Extract index template

* Remove renderer_for helper method

Now that we've extracted all pages, this has no references.

* Remove unused view factory class/config

This predates Arbre builder method which is how we create components. This may have been used to override views entirely but with partials we provide a simpler, more Rails like way to customize the admin while minimizing the library footprint and maintainance.

Note that we use the method_or_proc_helper throughout but mysteriously this is defined in view helpers but at the root level module. We already require and include the module elsewhere, e.g. menu item and scopes, so requiring it in filters/forms isn't abnormal. With the view factory removal, the test suite starting failing here but now all green.

* Remove unused main_content locale

* Reorganize some templates post extraction

This moves sidebar to top level. The action item and sidebar block templates now go into "shared" directory. This also removes an unnecessary head partial since we'll just include the html head defaults directly into the layout file now that we have that extracted.