Fix logic for user initiated dark mode toggle
If the system preference was used and it happen to be light mode and the user had not used this switch, then when toggling for the first time nothing would change because it would set the theme key to light mode when it already was. This updates the logic to match what we do when setting the theme so it has the correct, expected behavior.