Inline more styles into extracted partials (#8178)
* Add pagination styles in templates

* Remove unnecessary id

* Update site_title config docs

* Inline page content styles with link-default class

With the extract template we can now inline the grid layout styles and use an explicit named class for setting a link default style in content area, this way it can also be easily removed if not desired.

* Update tailwind content paths

* Update stylesheet to match Rails default

* Clean up unused content

Remove commented template

Remove unused font size base variables

* Fix svgToTinyDataUri indentation

* Inline devise styles in devise templates

Apply link default style component to links partial to retain previous styling that was removed.

* Fix dark mode JS error on Devise templates

These elements won't exist, although they could, we still need to protect against the fact they may not be there. We can use JS optional chaining operator here.

* Fix logic for user initiated dark mode toggle

If the system preference was used and it happen to be light mode and the user had not used this switch, then when toggling for the first time nothing would change because it would set the theme key to light mode when it already was. This updates the logic to match what we do when setting the theme so it has the correct, expected behavior.

* Remove kaminari pagination locale key

This is the default already which is what we want to use.