Isolate ActiveAdmin namespace
With the base, page and resource controllers migrated to app folder for autoloading and fixing helper bug changes not reloading in the host app, I think this has to be used with `isolate_namespace ActiveAdmin` in the engine file for this to be a safe change. With the namespace isolation in place, I encounter an error though when loading the login form page where the URL helpers don't seem to be accessible from the sessions new template. The admin pages work fine.