Create sidebar_sections partial with inline style
This also renames the partial sidebar_section_block to sidebar_section, to follow same approach from elsewhere, e.g. action items.

This adds default vertical spacing similar to what we had before and can now be customized by the host app.

Since active-filters are implemented as a sidebar, we no longer need the CSS style to set spacing since it's now in the partial as a customizable default.