Vendor Flowbite again for importmap setup
Since we need to have Tailwind scan the Flowbite JS for classes used, we already use the activeadmin gem path in the Tailwind config file, so we may as well just vendor this JS to make it easier.

It's been problematic to pin Flowbite as the builds from the supported providers don't work. We receive different errors in testing. The Flowbite docs suggest Cloudflare and that works fine when importing that URL. The jspm.org API doesn't support it which is what importmap-rails uses. So we are just vendoring the same file we would have pinned.