Update attributes_table to use data attributes
To follow table_for, this now uses data attributes for metadata like column name so it can still be styled but not clash. This also changes the default class name to be attributes-table. No need for odd/even classes since CSS has long supported that query selector (started the year ActiveAdmin was created).