Additional template and style updates (#8181)
* Extract resource comments partial

This partial is for customizing the entire output of active_admin_comments_for method. This used to be a component but is now an Arbre::Element which outputs no root element where as component does. This way the user can control all that from the partial. We may want to consider extracting the comments resource to partials too so for now this is named as resource_comments to avoid confusion.

* Inline styles in resource_comments partial

* Add empty-value style

* Update disabled styles for dropdown toggle

* Rename partial to action_item

* Create sidebar_sections partial with inline style

This also renames the partial sidebar_section_block to sidebar_section, to follow same approach from elsewhere, e.g. action items.

This adds default vertical spacing similar to what we had before and can now be customized by the host app.

Since active-filters are implemented as a sidebar, we no longer need the CSS style to set spacing since it's now in the partial as a customizable default.

* Lower vertical spacing in active filters list

* Use resource consistently in examples

* Add @rails/ujs to package json template

* Update table_for with data attributes for metadata

When ActiveAdmin was created, HTML data attributes and CSS even/odd selection was a new feature in some browsers that year. It was common to use class names for metadata since there wasn't any other means. This now removes any default class names so users can safely specify any Tailwind classes and rely on data attributes when needing to customize further based on that through plain CSS.

* Simplify data-table class name default

No need to manually change options like this since we can just use Arbre to add the default class name.

* Remove index_column from table_for

* Fix coverage patch for id_column edit link

We mostly just test the default actions so this adds coverage for the edit branch of the id_column method.

* Use modifier to hide/show dark mode icons

No need to query for the SVG elements since we can just use the standard Tailwind modifiers to control which is visible from the HTML. Defaults to light mode.

* Inline action item block render

No need for this to be in an Arbre partial. We only document link_to usage and we have noted in the upgrading guide to not use resource named methods here (e.g. post) and use the resource helper instead.

* Update status_tag with data-status attribute

Add the status given as a data attribute instead of a class in the same format as before (underscore). This way the only class set by default is the status-tag class. Caller can provide any class they want.

* Update attributes_table to use data attributes

To follow table_for, this now uses data attributes for metadata like column name so it can still be styled but not clash. This also changes the default class name to be attributes-table. No need for odd/even classes since CSS has long supported that query selector (started the year ActiveAdmin was created).

* Extract blank slate/empty results to partials

* Remove public non-action method

This also has no references