Fix app helper autoloading in dev and isolate namespace (#8180)
* Migrate base controller

* Update base_controller.rb

* Migrate page controller

* Migrate resource controller

* Update resource_controller.rb

* Remove autoload's

* Isolate ActiveAdmin namespace

With the base, page and resource controllers migrated to app folder for autoloading and fixing helper bug changes not reloading in the host app, I think this has to be used with `isolate_namespace ActiveAdmin` in the engine file for this to be a safe change. With the namespace isolation in place, I encounter an error though when loading the login form page where the URL helpers don't seem to be accessible from the sessions new template. The admin pages work fine.

* Use main_app proxy for Devise URL helpers

Now with isolate_namespace being used, these helpers will fail unless we specify main_app proxy per the Rails Engines guide.

* Update comment input name

This also changed with the isolate_namespace usage.