Refactor download format links helper (#8185)
This is undocumented. No need to provide since the formats can be set through the download_links config which is preferred. We'll update our config initializer template to note that the format must be registered with mime type if it isn't.

Eventually, we'll remove the XML and JSON defaults along with the respond_to for security reasons since they are still accessible as public routes even without links. Users should declare this for each resource where they need it.