Update dark mode styles
Still some elements to update with better dark mode styles but this gets most of the contents on the page.

Updated the main navigation layout and hover/selected states.

Updated the data-table and attributes-table styles, not just dark mode but spacing, borders and other styles.

Imrpoved the empty-value style for all modes too.