Replace deps with devDeps in package.json
By using importmap-rails we no longer need an NPM package with these dependencies so we can remove them. It seems that they would need to be listed on the app host package.json file anyway from testing. We'll stick to serving the JS through importmap-rails as the only option.

Due to how we generate the dev/test apps for gem development, we need to specify tailwindcss as a devDependency in package.json since the config file will point to `../../` the project root.