Use importmap-rails for JS assets (#8186)
* Use importmap-rails for JS

* Vendor Flowbite again for importmap setup

Since we need to have Tailwind scan the Flowbite JS for classes used, we already use the activeadmin gem path in the Tailwind config file, so we may as well just vendor this JS to make it easier.

It's been problematic to pin Flowbite as the builds from the supported providers don't work. We receive different errors in testing. The Flowbite docs suggest Cloudflare and that works fine when importing that URL. The jspm.org API doesn't support it which is what importmap-rails uses. So we are just vendoring the same file we would have pinned.

* Replace deps with devDeps in package.json

By using importmap-rails we no longer need an NPM package with these dependencies so we can remove them. It seems that they would need to be listed on the app host package.json file anyway from testing. We'll stick to serving the JS through importmap-rails as the only option.

Due to how we generate the dev/test apps for gem development, we need to specify tailwindcss as a devDependency in package.json since the config file will point to `../../` the project root.

* Replace jsbundling-rails with importmap-rails

Using importmap-rails will be easier for us to provide the JS for host app. With feedback we can see what we need to support jsbundling-rails without requiring a dependency on either gem. For tests and CI, we'll only use the importmap-rails approach.