Fix bug report workflow with importmap-rails
This adds importmap-rails as part of the bug report template script since its now necessary for it to run. Due to same issue we had before even with sassc, since we have another place where we reference `app.config.assets` in the engine file, we have to guard against it not being present for this bug report template script.

The workflow paths config has been updated to include Gemfile and gemspec to pick up future changes as typically that causes this to fail.

Since we removed sassc, the bug report template workflow now runs in about 30 seconds so ideally we could just always run this workflow on any PR and not just for specific changes.