Move kaminari views within the active_admin folder
Since it will be expected for users to extract many partials in v4, it would help to have all templates under the active_admin folder. Previously, we had Kaminarin in the default location but Kaminari has a views_prefix option we can set.

https://github.com/kaminari/kaminari#multiple-templates

With this that means the our custom templates can now live in app/views/active_admin/kaminari and be within the active_admin folder along side all other templates that can be extracted.