Update resource generator
This inlines the previous boilerplate class into the generator since no need to hide this behind a flag. We want the generated boilerplate to be the default. It should be active (uncommented) so it can be tried out and modified. We want users to rely on declaring their resource pages rather than rely on defaults-rendering.