Template and styles clean up (#8197)
* Remove date picker input classes

These are no longer needed and have been replaced by the native HTML date input. The DatepickerInput was for jQuery UI but has since been removed. Formtastic already supports specifying `input_html` which will allow users to use a custom date picker if they want.

* Add action-item-button class

* Add sorted icon for data-table

* Docs: Update has_many examples

* Update has-many form styles

* Scope empty-value to attributes-table

Since it was only used here. It was previously in the resource comments before partial extraction but no need there since the partial is meant to be copied and modified by host app.

* Remove empty orm/mongoid file

This doesn't seem to have ever been completed but mongo should still be usable. We have introduced other ActiveRecord error handling so may be unavoidable.

* Inline plural many count value

* Update resource generator

This inlines the previous boilerplate class into the generator since no need to hide this behind a flag. We want the generated boilerplate to be the default. It should be active (uncommented) so it can be tried out and modified. We want users to rely on declaring their resource pages rather than rely on defaults-rendering.

* RTL updates on flash msg and dark mode icons

* Simplify main menu JS and apply RTL support

Originally, tried this but had some trouble. When applying RTL support for the nested menu icon, I looked up using group with data attribute and found this: https://github.com/tailwindlabs/tailwindcss/discussions/11768#discussioncomment-6663339 With this applied the group class approach was working. This is ideal since now our JS is just toggling a simple data attribute and has no classes to set/remove. This also works with the RTL modifier. No need for additional Ruby here to set the class if its current, since we can just base it off the data attribute now.

* Remove leading/tracking styles

* Add comment on destroy redirect trade off

* Include commented out PunditAdapter

* Create tag_policy.rb

We need this in templates because when generating a dev app (rails_templates_with_data) if we enable authorization adapter for testing we receive errors regarding this model. Test suite is unaffected by this change.