Fix site_title helper for namespace and Devise
In Devise templates, the active_admin_namespace helper won't be available so we need to use active_admin_application but in the admin itself, we have to use the namespace to support site title that can be configured by namespace.

We didn't have tests beforehand, even previous release, to cover this case so this adds a feature spec that was failing pre-fix and passing post-fix.