Rename variable to activeAdminPath for clarity
Should help on the app host side since "activeAdmin" in the name to indicate that its scanning the files in the gem. Since there are complimentary ActiveAdmin gems, for sure those will update and need to be included in this file in the host app so they can follow a convention/pattern.