Migrate collection helpers to app/helpers
Some uses already rely on the `collection` helper method (InheritedResources::Base) as a default argument but others through a provided argument which happens to be the same name (!). We only use these 2 methods in views so we can define as a plain Rails helper.

Since we have a feature spec using the paginated_collection in a page, we have to include the helper in our base controller so it's available to both resources and pages.